device: ip_state_other should not block the connection if may-fail
When the device is configure to use dhcp/autoconf, the state must be
expanded to ip_state. In addition, if one of the IP configuration is
ready and the other may fail, then we should consider it ready. The
other ip state does not matter at all, it can be none too.

Fixes-test: @nmcli_device_reapply_routes