libnm: pass more parameters to _property_direct_set_string() in NMSetting
Both callers themselves needed to call _nm_setting_get_private(),
only to pass it to _property_direct_set_string().

Instead, pass the necessary parameters to _property_direct_set_string(),
so it can do that itself.

This additional parameters will be necessary when we add a hook for
setting the string.